% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkuData.R
\name{AkuData}
\alias{AkuData}
\title{Generate test data}
\usage{
AkuData(n)
}
\arguments{
\item{n}{Number of observations within each quarter.}
}
\value{
A data frame with the following variables: \item{id}{Sample unit
identifier} \item{year}{Year} \item{q}{Quarter} \item{month}{Month}
\item{R}{Response indicator} \item{age}{Age group} \item{sex}{Education
group} \item{famid}{Family identifier} \item{unemployed}{Unemployed}
\item{workforce}{In workforce}
}
\description{
Generate test data of eight quarters
}
\examples{

# Generates data - two years
z = AkuData(3000) # 3000 in each quarter

}
