% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateSSB.R
\name{CrossStrata}
\alias{CrossStrata}
\title{Crossing several factor variables}
\usage{
CrossStrata(by, sep = "-", returnb = FALSE, asNumeric = FALSE,
  byExtra = NULL)
}
\arguments{
\item{by}{Dataframe or matrix with several variables}

\item{sep}{Used to create new level names}

\item{returnb}{When TRUE an overview of original variabels according to new
levels are also retuned.}

\item{asNumeric}{When TRUE the new variable is numeric.}

\item{byExtra}{Contains the same variables as by and represents another data
set.}
}
\value{
\item{a}{The new variable} \item{aExtra}{New variable according to
byExtra} \item{b}{Overview of original variabels according to new levels}
}
\description{
Create new factor variable by crossing levels in several variables
}
\examples{

CrossStrata(cbind(factor(rep(1:3,2)),c('A',rep('B',5)) ))

}
