% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAndGetTotal.R
\name{MatchVarNames}
\alias{MatchVarNames}
\title{MatchVarNames}
\usage{
MatchVarNames(x, y, sep = ":", makeWarning = FALSE)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{sep}{sep}

\item{makeWarning}{Warning when matching by reordering}
}
\value{
An integer vector giving the position in y of the first match if there is a match, otherwise NA.
}
\description{
MatchVarNames
}
\examples{
z <- data.frame(A = factor(c("a", "b", "c")), B = factor(1:2), C = 1:6)
x <- colnames(model.matrix(~B * C * A, z))
y <- colnames(model.matrix(~A * B + A:B:C, z))
MatchVarNames(x, y)
}
\keyword{internal}
