% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAndGetTotal.R
\name{OrderedVarNames}
\alias{OrderedVarNames}
\title{OrderedVarNames}
\usage{
OrderedVarNames(x, sep = ":")
}
\arguments{
\item{x}{input}

\item{sep}{}
}
\value{
output
}
\description{
OrderedVarNames
}
\examples{
z <- data.frame(A = factor(c("a", "b", "c")), B = factor(1:2), C = 1:6)
x <- colnames(model.matrix(~B * C * A, z))
OrderedVarNames(x)
}
\keyword{internal}
