% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calSSB.R
\name{print.calSSB}
\alias{print.calSSB}
\title{Print method for calSSB}
\usage{
\method{print}{calSSB}(x, digits = max(getOption("digits") - 3, 3), ...)
}
\arguments{
\item{x}{calSSB object}

\item{digits}{positive integer.  Minimum number of significant digits to be used for printing most numbers.}

\item{\dots}{further arguments sent to the underlying}
}
\value{
Invisibly returns the original object.
}
\description{
Print method for calSSB
}
\keyword{print}
