% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calSSB.R
\name{print.calSSBwide}
\alias{print.calSSBwide}
\title{Print method for calSSBwide}
\usage{
\method{print}{calSSBwide}(x, digits = max(getOption("digits") - 3, 3), ...)
}
\arguments{
\item{x}{calSSBwide object}

\item{digits}{positive integer.  Minimum number of significant digits to be used for printing most numbers.}

\item{\dots}{further arguments sent to the underlying}
}
\value{
Invisibly returns the original object.
}
\description{
Print method for calSSBwide
}
\keyword{print}
