% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateSSB.R
\name{WideFromCalibrate}
\alias{WideFromCalibrate}
\title{Rearrange output from CalibrateSSB (calSSB object). Ready for input to PanelEstimation.}
\usage{
WideFromCalibrate(a, wave = NULL, id = NULL, subSet = NULL, extra = NULL)
}
\arguments{
\item{a}{A calSSB object. That is, output from CalibrateSSB() or CalSSBobj().}

\item{wave}{Time or another repeat variable.}

\item{id}{Identifier variable.}

\item{subSet}{Grouping variable for splitting ouput.}

\item{extra}{Dataset with extra variables not in \code{a}.}
}
\value{
Output has the same elements (+ extra) as input (a), but rearranged.
When subSet is input otput is alist according to the subSet levels.
}
\description{
One row for each id and one column for each wave.
}
\details{
When wave, id or extra is NULL, corresponding elements in the input object (\code{a}) will be used if available,
}
\examples{

# See examples in PanelEstimation and CalSSBobj

}
\seealso{
\code{\link{CalibrateSSB}}, \code{\link{CalSSBobj}}, \code{\link{PanelEstimation}}.
}
